Fix compiler and linker warnings
- Add type casts where C++20 deprecates implicit enum/float/int
  conversions (rays color_space, rays-video video, reflex shape/style)
- Cast NSImageRep* to NSBitmapImageRep* in rays bitmap.mm loader
- Remove unused const get_data overload for WallShape in reflex
- Wrap not-implemented stubs (SoundPlayer::rewind, vibrate,
  Timers::find_timers) with #pragma to suppress -Wmissing-noreturn
- xot extconf: only add -lstdc++ on gcc; clang++ auto-links libc++
  and the explicit -lc++ produced an ld duplicate-library warning


[[7A053F39-54F2-4D0A-AF3F-814E0C8DE517]]

+++ src/color_space.cpp
@@ -133,7 +133,7 @@ namespace Rays
 	ColorSpace::has_alpha () const
 	{
 		return
-			(ALPHA_FIRST <= type_ && type_ <= ALPHA_LAST) ||
+			(ALPHA_FIRST <= (int) type_ && (int) type_ <= ALPHA_LAST) ||
 			type_ == RGBA_8888  || type_ == ARGB_8888  ||
 			type_ == BGRA_8888  || type_ == ABGR_8888  ||
 			type_ == RGBA_float || type_ == ARGB_float ||
